+++ example/convex/_generated/api.d.ts
@@ -20,11 +20,27 @@ declare const fullApi: ApiFromModules<{
   example: typeof example;
 }>;
 
+/**
+ * A utility for referencing Convex functions in your app's public API.
+ *
+ * Usage:
+ * ```js
+ * const myFunctionReference = api.myModule.myFunction;
+ * ```
+ */
 export declare const api: FilterApi<
   typeof fullApi,
   FunctionReference<any, "public">
 >;
 
+/**
+ * A utility for referencing Convex functions in your app's internal API.
+ *
+ * Usage:
+ * ```js
+ * const myFunctionReference = internal.myModule.myFunction;
+ * ```
+ */
 export declare const internal: FilterApi<
   typeof fullApi,
   FunctionReference<any, "internal">
@@ -32,4 +48,5 @@ export declare const internal: FilterApi<
 
 export declare const components: {
   quota: import("../../../src/component/_generated/component.js").ComponentApi<"quota">;
+  secondary: import("../../../src/component/_generated/component.js").ComponentApi<"secondary">;
 };