+++ src/component/_generated/api.ts
@@ -25,11 +25,27 @@ const fullApi: ApiFromModules<{
   validators: typeof validators;
 }> = anyApi as any;
 
+/**
+ * A utility for referencing Convex functions in your app's public API.
+ *
+ * Usage:
+ * ```js
+ * const myFunctionReference = api.myModule.myFunction;
+ * ```
+ */
 export const api: FilterApi<
   typeof fullApi,
   FunctionReference<any, "public">
 > = anyApi as any;
 
+/**
+ * A utility for referencing Convex functions in your app's internal API.
+ *
+ * Usage:
+ * ```js
+ * const myFunctionReference = internal.myModule.myFunction;
+ * ```
+ */
 export const internal: FilterApi<
   typeof fullApi,
   FunctionReference<any, "internal">